+++ FlashCap.Core/Internal/LibAVFoundation/LibAVFoundation.cs
@@ -362,101 +362,7 @@ internal static partial class NativeMethods_AVFoundation
             }
         }
 
-        public sealed class AVCaptureVideoDataOutput : AVCaptureOutput
-        {
-            public AVCaptureVideoDataOutput() : base(IntPtr.Zero, retain: false)
-            {
-                Init();
-            }
-
-            private void Init()
-            {
-
-                var videoAlloc = LibObjC.SendAndGetHandle(
-                    LibObjC.GetClass("AVCaptureVideoDataOutput"),
-                    LibObjC.GetSelector(LibObjC.AllocSelector));
-                
-                
-                Handle = LibObjC.SendAndGetHandle(
-                    videoAlloc,
-                    LibObjC.GetSelector("init"));
-            }
-
-            public unsafe int[] AvailableVideoCVPixelFormatTypes =>
-                LibCoreFoundation.CFArray.ToArray(
-                    LibObjC.SendAndGetHandle(
-                        Handle,
-                        LibObjC.GetSelector("availableVideoCVPixelFormatTypes")),
-                    static handle =>
-                    {
-                        int value;
-                        if (LibCoreFoundation.CFNumberGetValue(handle, LibCoreFoundation.CFNumberType.sInt32Type, &value))
-                            return value;
-                        throw new InvalidOperationException("The value contained by CFNumber cannot be read as 32-bit signed integer.");
-                    });
-
-            public bool AlwaysDiscardsLateVideoFrames
-            {
-                get =>
-                    LibObjC.SendAndGetBool(
-                        Handle,
-                        LibObjC.GetSelector("alwaysDiscardsLateVideoFrames"));
-                set =>
-                    LibObjC.SendNoResult(
-                        Handle,
-                        LibObjC.GetSelector("setAlwaysDiscardsLateVideoFrames:"),
-                        value);
-            }
-
-            public unsafe void SetPixelFormatType(int format)
-            {
-                format = 1111970369;
-                
-                //var number = LibCoreFoundation.CFNumberCreate(IntPtr.Zero, LibCoreFoundation.CFNumberType.sInt32Type, &format);
-                
-                var nsNumber = LibObjC.CreateNSNumber(format);
-                
-                LibCoreFoundation.CFRetain(nsNumber);
-                
-                //var keys = new IntPtr[] { LibCoreVideo.kCVPixelBufferPixelFormatTypeKey};
-                //var values = new IntPtr[] { nsNumber };
-
-                /*var dictionary = LibCoreFoundation.CFDictionaryCreate(
-                    IntPtr.Zero,
-                    keys,
-                    values,
-                    numValues: 1,
-                    LibCoreFoundation.kCFCopyStringDictionaryKeyCallBacks,
-                    LibCoreFoundation.kCFTypeDictionaryValueCallBacks);*/
-
-                IntPtr nsDictionaryClass = LibCoreFoundation.BasicClassesHelper.NSDictionary;
-                IntPtr dictSel =  LibObjC.GetSelector("dictionaryWithObject:forKey:");
-                
-               //LibCoreFoundation.CFRetain(dictionary);
-
-                IntPtr nsPixelFormatKey = LibCoreVideo.kCVPixelBufferPixelFormatTypeKey;
-               
-                IntPtr videoSettings =  LibObjC.SendAndGetHandle(nsDictionaryClass, dictSel, nsNumber, nsPixelFormatKey);
-                LibCoreFoundation.CFRetain(videoSettings);
-                IntPtr setVideoSettingsSel = LibObjC.GetSelector("setVideoSettings:");
-                
-                LibObjC.SendNoResult(
-                    Handle,
-                    setVideoSettingsSel,
-                    videoSettings);
-
-                LibCoreFoundation.CFRelease(videoSettings);
-                LibCoreFoundation.CFRelease(nsNumber);
-                //LibCoreFoundation.CFRelease(strTrue.Handle);
-            }
-
-            public void SetSampleBufferDelegate(AVCaptureVideoDataOutputSampleBuffer sampleBufferDelegate, LibCoreFoundation.DispatchQueue sampleBufferCallbackQueue) =>
-                LibObjC.SendNoResult(
-                    Handle,
-                    LibObjC.GetSelector("setSampleBufferDelegate:queue:"),
-                    sampleBufferDelegate.Handle,
-                    sampleBufferCallbackQueue.Handle);
-        }
+       
 
         public abstract class AVCaptureVideoDataOutputSampleBuffer : LibObjC.NSObject
         {
